data-center/docker/relay_mqtt_db/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
ciso8601==2.3.1
idna==3.6
influx-client==1.9.0
influxdb-client==1.39.0
paho-mqtt==1.6.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytool==4.0.0
PyYAML==6.0.1
reactivex==4.0.4
requests==2.31.0
simplejson==3.19.2
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
